add make-task.js and make-filter.js modules
